+++ src/Eawv.Service/Models/CountryModel.cs
@@ -1,0 +1,11 @@
+// (c) Copyright by Abraxas Informatik AG
+// For license information see LICENSE file
+
+namespace Eawv.Service.Models;
+
+public class CountryModel
+{
+    public string IsoId { get; set; } = string.Empty;
+
+    public string Description { get; set; } = string.Empty;
+}