+++ src/Eawv.Service/Models/CountryXmlRootModel.cs
@@ -1,0 +1,16 @@
+﻿// (c) Copyright by Abraxas Informatik AG
+// For license information see LICENSE file
+
+using System;
+using System.Xml.Serialization;
+
+namespace Eawv.Service.Models;
+
+[Serializable]
+[XmlRoot("CountryCollection")]
+public class CountryXmlRootModel
+{
+    [XmlArray("countries")]
+    [XmlArrayItem("country")]
+    public CountryXmlModel[] Country { get; set; }
+}